+++ xxGraphicD3D10.cpp
@@ -648,7 +648,7 @@ void xxUnmapBufferD3D10(uint64_t device, uint64_t buffer)
 //==============================================================================
 //  Texture
 //==============================================================================
-uint64_t xxCreateTextureD3D10(uint64_t device, int format, int width, int height, int depth, int mipmap, int array, void const* external)
+uint64_t xxCreateTextureD3D10(uint64_t device, uint64_t format, int width, int height, int depth, int mipmap, int array, void const* external)
 {
     ID3D10Device* d3dDevice = reinterpret_cast<ID3D10Device*>(device);
     if (d3dDevice == nullptr)
