+++ xxGraphicD3D6.cpp
@@ -586,7 +586,7 @@ void xxUnmapBufferD3D6(uint64_t device, uint64_t buffer)
 //==============================================================================
 //  Texture
 //==============================================================================
-uint64_t xxCreateTextureD3D6(uint64_t device, int format, int width, int height, int depth, int mipmap, int array, void const* external)
+uint64_t xxCreateTextureD3D6(uint64_t device, uint64_t format, int width, int height, int depth, int mipmap, int array, void const* external)
 {
     DDSURFACEDESC2 desc = {};
     desc.dwSize = sizeof(DDSURFACEDESC2);
