+++ xxGraphicGlide.cpp
@@ -311,7 +311,7 @@ void xxUnmapBufferGlide(uint64_t device, uint64_t buffer)
 //==============================================================================
 //  Texture
 //==============================================================================
-uint64_t xxCreateTextureGlide(uint64_t device, int format, int width, int height, int depth, int mipmap, int array, void const* external)
+uint64_t xxCreateTextureGlide(uint64_t device, uint64_t format, int width, int height, int depth, int mipmap, int array, void const* external)
 {
     GrTexture* info = xxAlloc(GrTexture);
     if (info == nullptr)
