+++ xxGraphicVulkan.cpp
@@ -1485,7 +1485,7 @@ void xxUnmapBufferVulkan(uint64_t device, uint64_t buffer)
 //==============================================================================
 //  Texture
 //==============================================================================
-uint64_t xxCreateTextureVulkan(uint64_t device, int format, int width, int height, int depth, int mipmap, int array, void const* external)
+uint64_t xxCreateTextureVulkan(uint64_t device, uint64_t format, int width, int height, int depth, int mipmap, int array, void const* external)
 {
     VkDevice vkDevice = reinterpret_cast<VkDevice>(device);
     if (vkDevice == VK_NULL_HANDLE)
